renames connections container
