+++ src/entry.jsx
@@ -4,7 +4,7 @@ import createHashHistory from 'history/lib/createHashHistory'
 import { Router, Route } from 'react-router';
 import App from './containers/app.jsx';
 import configureStore from './store/configure';
-import ConnectionsContainer from './containers/connections.jsx';
+import ConnectionListContainer from './containers/connection-list.jsx';
 import DatabaseListContainer from './containers/database-list.jsx';
 
 const history = createHashHistory();
@@ -15,7 +15,7 @@ React.render(
     {() =>
       <Router history={history}>
         <Route component={App}>
-          <Route path="/" component={ConnectionsContainer} />
+          <Route path="/" component={ConnectionListContainer} />
           <Route path="/:id" component={DatabaseListContainer} />
         </Route>
       </Router>
